+++ src/model/library/niceNumbers.js
@@ -1,0 +1,3 @@
+export function numberToMoney(accountBallance) {
+  return accountBallance.toLocaleString('pt-BR', { style: 'currency', currency: 'BRL' });
+}