+++ src/model/library/utils.js
@@ -1,3 +1,0 @@
-export function getUserNameFromURL() {
-  return document.location.pathname.split("/")[2];
-}